refactor: Extract enums to models package and create filters.go
- Move CommandType and AttributeType enums to models/enums.go
- Create filters.go with FilterFunc and filter implementations
- Update imports in engine_nano.go and unified_commands.go
- Avoids circular dependencies while organizing related types
